рефакторинг админ-чата

+++ app/helpers.php
@@ -985,7 +985,7 @@ function statsChat()
  */
 function statsNewChat()
 {
-    return Chat::query()->max('created_at');
+    return (int) Chat::query()->max('created_at');
 }
 
 /**
